Rename missed files
