+++ ui/core-react/src/core-react/useTranslation.ts
@@ -5,15 +5,15 @@
 
 import { UiCore } from "./UiCore";
 
-interface Localization {
+interface UseTranslationResult {
   translate: (key: string) => string;
 }
 
 /**
- * Returns a localization object.
+ * Returns a translation function.
  * @internal
  */
-export function useTranslation(): Localization {
+export function useTranslation(): UseTranslationResult {
   const translate = (key: string) => {
     return UiCore.translate(key);
   };
