+++ ui/imodel-components-react/src/imodel-components-react/quantityformat/misc/ScientificType.tsx
@@ -11,7 +11,7 @@ import type { CommonProps } from "@itwin/core-react";
 import { ScientificType } from "@itwin/core-quantity";
 import type { SelectOption } from "@itwin/itwinui-react";
 import { Select } from "@itwin/itwinui-react";
-import { useTranslation } from "../../Translation";
+import { useTranslation } from "../../useTranslation";
 
 /** Properties of [[ScientificTypeSelector]] component.
  * @internal
