+++ ui/imodel-components-react/src/imodel-components-react/useTranslation.ts
@@ -5,15 +5,15 @@
 
 import { UiIModelComponents } from "./UiIModelComponents";
 
-interface Localization {
+interface UseTranslationResult {
   translate: (key: string) => string;
 }
 
 /**
- * Returns a localization object.
+ * Returns a translation function.
  * @internal
  */
-export function useTranslation(): Localization {
+export function useTranslation(): UseTranslationResult {
   const translate = (key: string) => {
     return UiIModelComponents.translate(key);
   };
